First upload
Uploading .csv file which contains all data points used for the project.
Additionally, uploading a Food class that can store all information in the .csv file and a function that will return a vector with all Food values.
The main file is incomplete and as of now is used to test and debug certain aspects of the code.